# api/openapi.yaml
openapi: 3.0.0
info:
  title: Library Management System API
  description: API for managing books, authors, and subscribers in a library
  version: 1.0.0
servers:
  - url: http://localhost:8081
paths:
...
  /search_books:
    get:
      summary: Search books
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookAuthorInfo'
  /search_authors:
    get:
      summary: Search authors
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorInfo'
...
  /login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
components:
  schemas:
...
    AuthorInfo:
      type: object
      properties:
        id:
          type: integer
        firstname:
          type: string
        lastname:
          type: string
        photo:
          type: string
    BookAuthorInfo:
      type: object
      properties:
        book_id:
          type: integer
        book_title:
          type: string
        author_id:
          type: integer
        book_photo:
          type: string
        is_borrowed:
          type: boolean
        book_details:
          type: string
        author_lastname:
          type: string
        author_firstname:
          type: string
...